Further separate out OS abstraction and make it configurable.
All actually GLib dependent code is moved to fluid_sys_glib.[ch] while
the more generic code remains directly in fluid_sys.[ch].

A new os-abstraction variable is added to the build system that
currently only supports "glib" and causes the GLib dependency to be
pulled in and the new files to be included.

Minimal #ifs are used to include the right headers depending on the
abstraction.
