+++ CMakeLists.txt
@@ -103,6 +103,8 @@ option ( enable-threads "enable multi-threading support (such as parallel voice 
 option ( enable-openmp "enable OpenMP support (parallelization of soundfont decoding, vectorization of voice mixing, etc.)" on )
 option ( enable-unicode "enable UNICODE build for Windows" on )
 
+set ( osal "glib" CACHE STRING "OS abstraction to use, provided by src/utils/fluid_sys_${osal}.*" )
+
 # Platform specific options
 if ( CMAKE_SYSTEM MATCHES "Linux" )
     option ( enable-systemd "compile systemd support (if it is available)" on )
@@ -535,13 +537,15 @@ if ( CMAKE_VERSION VERSION_GREATER_EQUAL 3.15 AND VCPKG_TOOLCHAIN )
   set ( CMAKE_FIND_PACKAGE_PREFER_CONFIG ON )
 endif ()
 
-# Mandatory libraries: glib and gthread
-find_package ( GLib2 ${GLIB2_MINUMUM_VERSION} REQUIRED )
-list( APPEND PC_REQUIRES_PRIV "glib-2.0" "gthread-2.0")
+if ( osal STREQUAL "glib" )
+    # Mandatory libraries: glib and gthread
+    find_package ( GLib2 ${GLIB2_MINUMUM_VERSION} REQUIRED )
+    list( APPEND PC_REQUIRES_PRIV "glib-2.0" "gthread-2.0")
 
-if ( GLib2_VERSION AND GLib2_VERSION VERSION_LESS "2.26.0" )
-    message ( WARNING "Your version of glib is very old. This may cause problems with fluidsynth's sample cache on Windows. Consider updating to glib 2.26 or newer!" )
-endif ( GLib2_VERSION AND GLib2_VERSION VERSION_LESS "2.26.0" )
+    if ( GLib2_VERSION AND GLib2_VERSION VERSION_LESS "2.26.0" )
+        message ( WARNING "Your version of glib is very old. This may cause problems with fluidsynth's sample cache on Windows. Consider updating to glib 2.26 or newer!" )
+    endif ( GLib2_VERSION AND GLib2_VERSION VERSION_LESS "2.26.0" )
+endif ( osal STREQUAL "glib" )
 
 # Optional features
 unset ( LIBSNDFILE_SUPPORT CACHE )
