+++ test/test_seq_event_queue_sort.c
@@ -2,7 +2,7 @@
 #include "test.h"
 #include "fluidsynth.h" // use local fluidsynth header
 #include "fluid_event.h"
-#include "fluidsynth_priv.h"
+#include "fluid_sys.h"
 
 #include <limits.h>
